refactor(lsp): split project import tests
